fix(03.1-04): fuzz-safe INT32 bounds on query params + pushover profile validation
- Reduce query/path parameter bounds from INT64_MAX to INT32_MAX across
  all 21 API route files to prevent ASGI transport OverflowError when
  Schemathesis generates boundary values (C int overflow at 2^63)
- Add INT32_MIN/INT32_MAX constants to schemas/constraints.py; body
  (JSON) fields retain full INT64 range
- Cap applied_threshold and interviewing_threshold to 3650 days (10yr)
  to prevent timedelta overflow in ghost detection
- Add ProfileNotFoundError + _validate_profile to pushover service so
  notification endpoints return 404 instead of 500 for missing profiles
- Add profile validation error handling to all pushover API routes
- Fix ApplicationUpdate status validator to reject non-string types
  (e.g. {} from Schemathesis) with clear ValueError
- Result: 139 fuzz endpoints pass, 1 xfail (datetime RFC 3339), zero 500s

+++ src/career_os/api/analytics.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 
 from career_os.database import get_db
 from career_os.schemas.analytics import AnalyticsResponse
-from career_os.schemas.constraints import INT64_MAX
+from career_os.schemas.constraints import INT32_MAX
 from career_os.services.analytics import get_analytics
 
 router = APIRouter(prefix="/api/analytics", tags=["analytics"])
@@ -15,7 +15,9 @@ router = APIRouter(prefix="/api/analytics", tags=["analytics"])
 
 @router.get("")
 async def analytics(
-    profile_id: Annotated[int, Query(ge=1, le=INT64_MAX, description="Profile to show analytics for")],
+    profile_id: Annotated[
+        int, Query(ge=1, le=INT32_MAX, description="Profile to show analytics for")
+    ],
     db: Annotated[Session, Depends(get_db)],
 ) -> AnalyticsResponse:
     """Get analytics dashboard data.
